8-26-23

+++ AutoTradeMobile/DataClasses/PercentConverter.cs
@@ -11,12 +11,12 @@ namespace AutoTradeMobile
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            return (double)value;
+            return (decimal)value;
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            return (double)value;
+            return (decimal)value;
         }
     }
